fix: fix editor formatting
